requirements.txt
docx
pandas
numpy
sklearn.linear_model
matplotlib.pyplot
datetime
os

